refactor(find_vectors): split 3131-line module into a package
Behaviour-preserving split of spyde/actions/find_vectors.py into a package:
  - gpu_runtime.py  numba-CUDA/CuPy infra: pools, streams, latches, policy
  - kernels.py      numba @cuda.jit kernels + CuPy/cuFFT NXCORR
  - detectors.py    per-frame cores (NXCORR, DoG, beam-stop, subpixel, intensity)
  - chunk.py        ghost-block chunk pipeline (GPU + CPU + DoG)
  - orchestrate.py  dask batch orchestration (_do_compute_vectors, _nav_chunk_size)

__init__.py re-exports the public surface so external import sites
(from spyde.actions.find_vectors import X) keep working. No logic changes;
the memory-safety contract in _do_compute_vectors and all GPU latch/warmup
state are moved verbatim.
